+++ docs/internal/processes/development-guide-relevant-files.md
@@ -151,6 +151,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/factory/event_history_test.go` plus sibling `event_history_*.go` test owners | Event-history generated payload and canonical stream regression seam | Keep initial-structure payload assertions, script-boundary stream coverage, and chaining-trace lineage regressions in same-package sibling test owners so the emitted event contract stays reviewable by concern without moving helpers into a separate test package. |
 | `pkg/interfaces/public_factory_enums.go` | Shared public factory enum normalizers | Shared worker and workstation alias ownership lives here; exported callers must choose explicit strict or permissive helpers instead of duplicating alias tables. |
 | `pkg/interfaces/factory_runtime.go` | Handwritten public work-request boundary structs | Watched-file batch ingestion, generated worker output parsing, and fixture helpers still marshal these structs directly, so their JSON tags must stay aligned with the camelCase OpenAPI contract. |
+| `pkg/interfaces/` | Shared backend contract package-size seam | Keep the root contract package under the file-count cap by consolidating tiny leaf contracts into stable owner files instead of splitting exported interfaces into dependency-heavy subpackages. Preserve filenames that are referenced by contract guards or maintainer docs, such as `runtime_lookup.go`, `factory_runtime.go`, `public_factory_enums.go`, `work_dispatch.go`, `safe_diagnostics.go`, and `world_state_clones.go`. |
 | `pkg/cli/dashboard/dashboard.go` plus sibling `pkg/cli/dashboard/dashboard_*.go` files | CLI dashboard session-summary boundary | Keep dashboard render entrypoints, render-data shaping, dispatch-history fallback traversal, and formatting helpers in the same package, but split them into sibling `dashboard_*.go` files when size cleanup is needed so CLI/dashboard work stays behavior-preserving without duplicating traversal or formatting policy in neighboring packages. |
 | `pkg/cli/dashboard/dashboard_test.go` | CLI dashboard behavioral regression surface | Package tests for dashboard rendering and session-summary mapping should assert observable labels and failed-work detail output, especially when `PlaceCategoriesByID` is absent and the CLI must fall back to dispatch history. |
 | `pkg/config/factory_config_mapping.go` | Generated factory-config boundary alias rejection owner | Mirrored legacy alias guards for top-level objects and nested `definition` objects live here; keep one field inventory per boundary type, vary only the path passed into the shared rejection helper, and stop the helper at the documented generated-boundary depth so cleanup work does not broaden alias rejection into deeper nested `definition` chains. |
