+++ pkg/interfaces/factory_config.go
@@ -5,6 +5,125 @@ import (
 	"time"
 )
 
+// File directories
+
+const (
+	FactoryDir      = "factory"
+	WorkflowsDir    = "workflows"
+	WorkTypesDir    = "work-types"
+	InputsDir       = "inputs"
+	WorkersDir      = "workers"
+	WorkstationsDir = "workstations"
+
+	DefaultChannelName = "default"
+	ArtifactsDirectory = "artifacts"
+	StateDir           = "state"
+	TransitionsDir     = "transitions"
+	ArcsDir            = "arcs"
+	GuardsDir          = "guards"
+	ResourcesDir       = "resources"
+	TokensDir          = "tokens"
+	MarkingsDir        = "markings"
+
+	FactoryAgentsFileName = "AGENTS.md"
+
+	// relevant files.
+	FactoryConfigFile         = "factory.json"
+	CurrentFactoryPointerFile = ".current-factory"
+	MetadataFile              = "metadata.json"
+	MarkingFile               = "marking.json"
+)
+
+// Extensions
+
+const (
+	JsonExtension = ".json"
+)
+
+// Token fields
+// These are fields that are used to color tokens as they go through the petri net.
+const (
+	WorkID     = "work_id"
+	WorkTypeID = "work_type_id"
+	TraceID    = "trace_id"
+	ParentID   = "parent_id"
+	Tags       = "tags"
+	Relations  = "relations"
+	Payload    = "payload"
+)
+
+// Internal time work contract.
+const (
+	SystemTimeWorkTypeID         = "__system_time"
+	SystemTimePendingState       = "pending"
+	SystemTimePendingPlaceID     = SystemTimeWorkTypeID + ":" + SystemTimePendingState
+	SystemTimeExpiryTransitionID = SystemTimeWorkTypeID + ":expire"
+
+	SystemTimeDashboardWorkTypeID         = "time"
+	SystemTimeDashboardPendingPlaceID     = SystemTimeDashboardWorkTypeID + ":" + SystemTimePendingState
+	SystemTimeDashboardExpiryTransitionID = SystemTimeDashboardWorkTypeID + ":expire"
+
+	TimeWorkTagKeySource          = "agent_factory.source"
+	TimeWorkTagKeyCronWorkstation = "agent_factory.cron.workstation"
+	TimeWorkTagKeyNominalAt       = "agent_factory.time.nominal_at"
+	TimeWorkTagKeyDueAt           = "agent_factory.time.due_at"
+	TimeWorkTagKeyExpiresAt       = "agent_factory.time.expires_at"
+	TimeWorkTagKeyJitter          = "agent_factory.time.jitter"
+
+	TimeWorkSourceCron = "cron"
+)
+
+// IsSystemTimeWorkType reports whether a work type is the internal time-work
+// type used for cron ticks.
+func IsSystemTimeWorkType(workTypeID string) bool {
+	return workTypeID == SystemTimeWorkTypeID
+}
+
+// IsSystemTimePlace reports whether a place belongs to the internal time-work
+// state machine.
+func IsSystemTimePlace(placeID string) bool {
+	return placeID == SystemTimePendingPlaceID
+}
+
+// IsSystemTimeToken reports whether a token is an internal time-work token.
+func IsSystemTimeToken(token *Token) bool {
+	return token != nil && IsSystemTimeWorkType(token.Color.WorkTypeID)
+}
+
+// Executor configuration
+// These are keys to store data for the messages between executors.
+const (
+	// These are well known tags that are used to store data for the messages between executors.
+	RejectionFeedback = "_rejection_feedback"
+)
+
+// Resource constants
+
+// Resource avialable states
+const (
+	ResourceStateAvailable = "available"
+)
+
+// WorkerType constants for worker AGENTS.md frontmatter.
+const (
+	WorkerTypeModel  = "MODEL_WORKER"
+	WorkerTypeScript = "SCRIPT_WORKER"
+	WorkerTypeHosted = "HOSTED_WORKER"
+)
+
+// Hosted worker provider constants for public hosted worker config.
+const (
+	HostedWorkerProviderLinear = "LINEAR"
+)
+
+// WorkstationType constants for workstation AGENTS.md frontmatter.
+const (
+	WorkstationTypeModel    = "MODEL_WORKSTATION"
+	WorkstationTypeInvoke   = "MODEL_INVOKE"
+	WorkstationTypeLogical  = "LOGICAL_MOVE"
+	WorkstationTypeClassify = "CLASSIFIER_WORKSTATION"
+)
+
 // FactoryConfig is the specification of a factory as a JSON file.
 type FactoryConfig struct {
 	Name             string                          `json:"name"`
@@ -324,4 +443,88 @@ type ResolvedModelOperationBinding struct {
 	Slot    string                      `json:"slot"`
 	Source  ModelOperationBindingSource `json:"source"`
 	Content []WorkContentPart           `json:"content,omitempty"`
+}
+
+// WorkerConfig is the canonical worker configuration used by factory.json,
+// worker AGENTS.md frontmatter, and loaded runtime config.
+type WorkerConfig struct {
+	Name             string                    `json:"name" yaml:"name,omitempty"`
+	Type             string                    `json:"type" yaml:"type"`
+	Provider         string                    `json:"provider,omitempty" yaml:"provider,omitempty"`
+	Model            string                    `json:"model,omitempty" yaml:"model,omitempty"`
+	ModelProvider    string                    `json:"modelProvider,omitempty" yaml:"modelProvider,omitempty"`
+	ModelLocality    string                    `json:"modelLocality,omitempty" yaml:"modelLocality,omitempty"`
+	ExecutorProvider string                    `json:"executorProvider,omitempty" yaml:"executorProvider,omitempty"`
+	Operations       []ModelOperation          `json:"operations,omitempty" yaml:"operations,omitempty"`
+	Command          string                    `json:"command,omitempty" yaml:"command,omitempty"`
+	Args             []string                  `json:"args,omitempty" yaml:"args,omitempty"`
+	Resources        []ResourceConfig          `json:"resources,omitempty" yaml:"resources,omitempty"`
+	Timeout          string                    `json:"timeout,omitempty" yaml:"timeout,omitempty"`
+	StopToken        string                    `json:"stopToken,omitempty" yaml:"stopToken,omitempty"`
+	SkipPermissions  bool                      `json:"skipPermissions,omitempty" yaml:"skipPermissions,omitempty"`
+	Auth             *HostedWorkerAuthConfig   `json:"auth,omitempty" yaml:"auth,omitempty"`
+	Linear           *HostedLinearWorkerConfig `json:"linear,omitempty" yaml:"linear,omitempty"`
+	Body             string                    `json:"body,omitempty" yaml:"-"`
+
+	// Internal-only runtime fields retained during contract cleanup.
+	SessionID   string `json:"-" yaml:"-"`
+	Concurrency int    `json:"-" yaml:"-"`
+}
+
+type HostedWorkerAuthConfig struct {
+	SecretRef string `json:"secretRef,omitempty" yaml:"secretRef,omitempty"`
+}
+
+type HostedLinearWorkerConfig struct {
+	PollInterval string                          `json:"pollInterval,omitempty" yaml:"pollInterval,omitempty"`
+	TeamIDs      []string                        `json:"teamIds,omitempty" yaml:"teamIds,omitempty"`
+	StateIDs     []string                        `json:"stateIds,omitempty" yaml:"stateIds,omitempty"`
+	Mapping      HostedLinearWorkerMappingConfig `json:"mapping,omitempty" yaml:"mapping,omitempty"`
+	Claim        *HostedLinearWorkerClaimConfig  `json:"claim,omitempty" yaml:"claim,omitempty"`
+}
+
+type HostedLinearWorkerMappingConfig struct {
+	WorkType string `json:"workType,omitempty" yaml:"workType,omitempty"`
+	State    string `json:"state,omitempty" yaml:"state,omitempty"`
+}
+
+type HostedLinearWorkerClaimConfig struct {
+	AssigneeField string `json:"assigneeField,omitempty" yaml:"assigneeField,omitempty"`
+}
+
+// TimeoutDuration parses Timeout as a time.Duration. It returns zero when the
+// value is empty or invalid.
+func (w *WorkerConfig) TimeoutDuration() time.Duration {
+	if w.Timeout == "" {
+		return 0
+	}
+	d, _ := time.ParseDuration(w.Timeout)
+	return d
+}
+
+const (
+	ModelLocalityLocal = "LOCAL"
+	ModelLocalityCloud = "CLOUD"
+)
+
+const (
+	ModelOperationContentTypeText   = "TEXT"
+	ModelOperationContentTypeImage  = "IMAGE"
+	ModelOperationContentTypeAudio  = "AUDIO"
+	ModelOperationContentTypeJSON   = "JSON"
+	ModelOperationContentTypeBinary = "BINARY"
+)
+
+// ModelOperation declares one provider-agnostic capability exposed by a model worker.
+type ModelOperation struct {
+	Name    string               `json:"name" yaml:"name"`
+	Inputs  []ModelOperationSlot `json:"inputs,omitempty" yaml:"inputs,omitempty"`
+	Outputs []ModelOperationSlot `json:"outputs,omitempty" yaml:"outputs,omitempty"`
+}
+
+// ModelOperationSlot declares one named operation input or output slot.
+type ModelOperationSlot struct {
+	Name         string   `json:"name" yaml:"name"`
+	ContentTypes []string `json:"contentTypes,omitempty" yaml:"contentTypes,omitempty"`
+	Required     bool     `json:"required,omitempty" yaml:"required,omitempty"`
 }
