+++ pkg/interfaces/factory_events.go
@@ -1,6 +1,8 @@
 package interfaces
 
-import factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+import (
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+)
 
 // FactoryEventStream carries replayed history and then live canonical events.
 type FactoryEventStream struct {
