+++ pkg/interfaces/work_execution.go
@@ -1,6 +1,10 @@
 package interfaces
 
-import "time"
+import (
+	"time"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+)
 
 // WorkerState is a point-in-time snapshot of the dispatcher's state.
 type WorkerState struct {
@@ -108,4 +112,114 @@ type CommandDiagnostic struct {
 type PanicDiagnostic struct {
 	Message string `json:"message,omitempty"`
 	Stack   string `json:"stack,omitempty"`
+}
+
+// WorkFailureFamily captures the runtime behavior category for a normalized
+// work failure.
+type WorkFailureFamily string
+
+const (
+	WorkFailureFamilyTerminal  WorkFailureFamily = "terminal"
+	WorkFailureFamilyRetryable WorkFailureFamily = "retryable"
+	WorkFailureFamilyThrottle  WorkFailureFamily = "throttle"
+)
+
+// ProviderErrorFamily remains as a compatibility alias while runtime paths
+// transition to generalized work-failure naming.
+type ProviderErrorFamily = WorkFailureFamily
+
+const (
+	ProviderErrorFamilyTerminal  ProviderErrorFamily = WorkFailureFamilyTerminal
+	ProviderErrorFamilyRetryable ProviderErrorFamily = WorkFailureFamilyRetryable
+	ProviderErrorFamilyThrottle  ProviderErrorFamily = WorkFailureFamilyThrottle
+)
+
+// WorkFailureType is the stable customer-facing normalized failure type for
+// scoped runtime work execution paths.
+type WorkFailureType string
+
+const (
+	WorkFailureTypeAuthFailure         WorkFailureType = "auth_failure"
+	WorkFailureTypePermanentBadRequest WorkFailureType = "permanent_bad_request"
+	WorkFailureTypeThrottled           WorkFailureType = "throttled"
+	WorkFailureTypeInternalServerError WorkFailureType = "internal_server_error"
+	WorkFailureTypeTimeout             WorkFailureType = "timeout"
+	WorkFailureTypeUnknown             WorkFailureType = "unknown"
+	WorkFailureTypeMisconfigured       WorkFailureType = "misconfigured"
+)
+
+// ProviderErrorType remains as a compatibility alias while runtime paths
+// transition to generalized work-failure naming.
+type ProviderErrorType = WorkFailureType
+
+const (
+	ProviderErrorTypeAuthFailure         ProviderErrorType = WorkFailureTypeAuthFailure
+	ProviderErrorTypePermanentBadRequest ProviderErrorType = WorkFailureTypePermanentBadRequest
+	ProviderErrorTypeThrottled           ProviderErrorType = WorkFailureTypeThrottled
+	ProviderErrorTypeInternalServerError ProviderErrorType = WorkFailureTypeInternalServerError
+	ProviderErrorTypeTimeout             ProviderErrorType = WorkFailureTypeTimeout
+	ProviderErrorTypeUnknown             ProviderErrorType = WorkFailureTypeUnknown
+	ProviderErrorTypeMisconfigured       ProviderErrorType = WorkFailureTypeMisconfigured
+)
+
+// WorkFailureDecision is the normalized behavior contract consumed by
+// downstream retry, termination, and throttle-pause logic.
+type WorkFailureDecision struct {
+	Retryable             bool
+	Terminal              bool
+	TriggersThrottlePause bool
+}
+
+// ProviderFailureDecision remains as a compatibility alias while runtime paths
+// transition to generalized work-failure naming.
+type ProviderFailureDecision = WorkFailureDecision
+
+// WorkFailureMetadata carries the normalized failure contract
+// across runtime boundaries after the original error has been rendered.
+type WorkFailureMetadata struct {
+	Family WorkFailureFamily `json:"family"`
+	Type   WorkFailureType   `json:"type"`
+}
+
+// ProviderFailureMetadata remains as a compatibility alias while runtime paths
+// transition to generalized work-failure naming.
+type ProviderFailureMetadata = WorkFailureMetadata
+
+// CanonicalWorkFailureMetadata returns the generalized failure metadata from
+// the runtime result, falling back to the legacy provider-named field while
+// older callers are still being migrated.
+func CanonicalWorkFailureMetadata(failure *WorkFailureMetadata, providerFailure *ProviderFailureMetadata) *WorkFailureMetadata {
+	if failure != nil {
+		return failure
+	}
+	return providerFailure
+}
+
+// ReplayArtifact is the versioned, self-contained recording used to replay a
+// factory run without requiring the original customer files or live side
+// effects.
+type ReplayArtifact struct {
+	SchemaVersion string                    `json:"schemaVersion"`
+	RecordedAt    time.Time                 `json:"recordedAt"`
+	Events        []factoryapi.FactoryEvent `json:"events"`
+
+	// The fields below are hydrated from Events for the current replay
+	// implementation. They are intentionally excluded from artifact storage.
+	Factory     factoryapi.Factory       `json:"-"`
+	Diagnostics ReplayDiagnostics        `json:"-"`
+	WallClock   *ReplayWallClockMetadata `json:"-"`
+}
+
+// ReplayDiagnostics stores artifact-level notes and optional nested execution
+// details.
+type ReplayDiagnostics struct {
+	Notes   []string                       `json:"notes,omitempty"`
+	Workers map[string]SafeWorkDiagnostics `json:"workers,omitempty"`
+}
+
+// ReplayWallClockMetadata retains wall-clock timing for investigation only.
+// Replay behavior is driven by logical ticks, not these timestamps.
+type ReplayWallClockMetadata struct {
+	StartedAt  time.Time `json:"started_at,omitempty"`
+	FinishedAt time.Time `json:"finished_at,omitempty"`
 }
